test(pg): the get-entry tool test requires the test database instead of skipping (#878)
The live suite in src/tools/__tests__/get-entry.test.ts read
OPENBRAIN_TEST_DATABASE_URL itself and swapped in describe.skip when the
variable was unset, so a run without a database reported 0 pass / 2 skip and
exited 0 -- indistinguishable at the exit code from a suite that ran.

Split by subject: the mock-pool suite stays in get-entry.test.ts, and the live
suite moves verbatim into a new src/tools/__tests__/get-entry.pg.test.ts as a
plain describe with a module-scope pool built from requireTestDatabaseUrl().
Absent the variable the helper throws test_database_required and the run fails
loudly. Test bodies, names, and assertions are unchanged.

The pre-commit gate lints staged files whole, and the remaining mock-pool
describe carried a pre-existing max-lines-per-function finding (410 lines) that
survived the move. .oxlintrc.json declines that exemption on purpose -- the debt
is paid per file as tests are touched -- so the mock suite is split by subject
into six describes, each under the rule value. No test body, name, or assertion
changed; only the describe wrappers they sit in. These are mock suites, so no
assert-db-tests-ran entry names them.

scripts/assert-db-tests-ran.ts registers the suite (minTests 1) and the global
floor moves 265 -> 266 so CI's anti-skip guard notices if it ever vanishes.

Done-means: scripts/done-means/878-pg-tests-require-database.sh
